Some refinement of the Application concept
