+++ src/lib.rs
@@ -7,6 +7,9 @@ pub mod inject;
 /// Configuration utilities based on Figment
 pub mod config;
 
+/// Application initialization
+pub mod app;
+
 pub use config::loader::{Config, ConfigLoader};
 pub use inject::{
     provide, to_provider_error, Error as InjectError, Inject, Provider, Result as InjectResult, Tag,
